[!!!][FEATURE] Proportional scroll handle/knob
This patch adds two options, `proportional` (boolean)
and `proportionalMinHeight` (integer) in order to make
the scrollbar's knob proportional to its content like
it's used to be in common GUIs.

Because this new option now defaults to true, it's
considered a breaking change [!!!] because people might
not expect it. Closes #7.

Besides, this patch closes #15 correcting an incorrect
use of mouseenter and mouseover.